Slice 10 task 1: the permission set at full width, and the half of the task the roadmap forbids
`LotTable.Zone` is a `ushort`, so `CommandKind.Zone`'s authored permission
set reaches the Lot whole. That discharges the narrowing `Command.cs` has
recorded against itself since slice 5, which cast a 16-bit set down to a
byte on the way in.

**Re-baseline, deliberate.** A saved column changed width, so the fold sees
twice the bytes for that column and every hash containing a Lot moves. Both
golden artefacts are re-recorded and nothing else in the suite moved, which
is the signature this change should have: `05 §4` says a different
permission set is a different city, and this is that. The fold, the
composition order and `Randomness.Mix` are untouched, so `World.HashSeed`
keeps its version byte per README step 3.

**The assertion nobody had written is why the narrowing survived five
slices.** `A_zone_command_paints_a_lot` checked that a Lot appeared where
the command said and checked nothing about what was painted on it, so the
cast passed every test in the file. The new test paints
`0b1000_0001_0000_0101` — four bits, two of them above a byte, because a
single bit cannot tell a set apart from an enum that happens to fit.

**The other half of task 1 is cut, and the reason is an ordering rather
than a difficulty.** The task read *"and a `zone` verb that paints an
area"*. `02 §2.2` says Lots are **generated, not painted** — a subdivider
carves zoned land against the Street network and refuses frontage-less
land — and `06` puts the Road Graph and Streets at milestone **5a**, which
is Phase 2. So the roadmap placed the subdivider after the slice meant to
make this verb honest. Painting a region here would stand in for more of 5a
than painting one Lot does, and every Lot it invented would be one the real
subdivider would have refused.

Nothing downstream wanted it: `SyntheticCity` already makes 225 Lots per
1,000 Citizens, a Zone in this build is not an entity but the set of Lots
carrying a permission bit, and task 9's Zone-size sweep samples that. The
verb keeps its single Lot and the generator stays visibly unbuilt.

Two entries filed to `0012` rather than to `0002`, because they are
corrections and not questions: **the Lot subdivider has no milestone
anywhere** — 5a names the graph and the geometry risk, the no-milestone
table omits it, and `adr/0014` and `adr/0035` both reason from its frontage
rule — and **`02 §2.2`'s opening sentence is false of every world this
project has run**, since every Lot in the tree is painted. Related: *every
Building is on the Road Graph by construction*, which `CONTEXT` → Frontage
leans on to delete the utility network, is currently true by there being no
Road Graph rather than by construction.

+++ src/Borough.Core/Entities/LotTable.cs
@@ -31,7 +31,7 @@ public sealed class LotTable
 
         East = _rows.Saved<Tiles>("east");
         North = _rows.Saved<Tiles>("north");
-        Zone = _rows.Saved<byte>("zone");
+        Zone = _rows.Saved<ushort>("zone");
 
         _rows.Seal();
     }
@@ -45,11 +45,35 @@ public sealed class LotTable
     /// <summary>Position along the north axis, in whole Tiles.</summary>
     public Column<Tiles> North { get; }
 
-    /// <summary>Which Zone family the Lot is zoned for. Resolved through the Ruleset.</summary>
-    public Column<byte> Zone { get; }
+    /// <summary>
+    /// The Lot's Zone: <b>a permission set, one bit per kind admitted here</b>.
+    /// </summary>
+    /// <remarks>
+    /// <para>
+    /// <b>A set rather than a kind, and the distinction is the design's rather than a convenience.</b>
+    /// <c>CONTEXT</c> → Zone is <em>"a permission set over land: it lists the uses allowed there and
+    /// forbids every other"</em>, and says mixed use needs no machinery because it is a set with more
+    /// than one entry. A single enum here would re-introduce the *zone type* that framing exists to
+    /// refuse, and would make mixed use something somebody has to add later.
+    /// </para>
+    /// <para>
+    /// <b>It is permission and never instruction</b> (<c>adr/0025</c>). Zoning admits a kind; it does
+    /// not summon one, and a bit set over land nothing wants to build on grows nothing. Density is the
+    /// intensity cap <em>within</em> a permission rather than a second concept.
+    /// </para>
+    /// <para>
+    /// <b>Sixteen bits, matching <see cref="Input.Command.Zone"/> at full width</b>, which discharges
+    /// the narrowing that verb has carried since slice 5 — it authored a set and this column kept a
+    /// byte of it. Sixteen is therefore how many kinds can ever be zoned for, against a <c>kind</c>
+    /// that is a <see cref="byte"/> everywhere else. The two are deliberately not the same width: a
+    /// kind nothing zones for is an ordinary thing — a service Building is *placed* (<c>adr/0032</c>)
+    /// — and a seventeenth zonable kind is a widening that should be argued rather than absorbed.
+    /// </para>
+    /// </remarks>
+    public Column<ushort> Zone { get; }
 
     /// <summary>Allocates a Lot at a position.</summary>
-    public Handle<Lot> Create(Tiles east, Tiles north, byte zone)
+    public Handle<Lot> Create(Tiles east, Tiles north, ushort zone)
     {
         Handle<Lot> handle = _rows.Allocate();
         int slot = _rows.Resolve(handle);
